refactor: Address 도메인 분리
