+++ be/src/main/java/kr/codesquad/secondhand/api/address/repository/AddressRepositoryImpl.java
@@ -1,6 +1,6 @@
-package kr.codesquad.secondhand.api.member.repository;
+package kr.codesquad.secondhand.api.address.repository;
 
-import kr.codesquad.secondhand.api.member.domain.Address;
+import kr.codesquad.secondhand.api.address.domain.Address;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
