+++ be/src/main/java/kr/codesquad/secondhand/api/member/dto/AddressSliceResponse.java
@@ -2,7 +2,7 @@ package kr.codesquad.secondhand.api.member.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kr.codesquad.secondhand.api.member.domain.Address;
+import kr.codesquad.secondhand.api.address.domain.Address;
 import lombok.Getter;
 
 @Getter
@@ -23,17 +23,17 @@ public class AddressSliceResponse {
     }
 
     @Getter
-    public static class AddressesResponse {
+    private static class AddressesResponse {
 
-        private Long id;
-        private String name;
+        private final Long id;
+        private final String name;
 
-        public AddressesResponse(Long id, String name) {
+        private AddressesResponse(Long id, String name) {
             this.id = id;
             this.name = name;
         }
 
-        public static AddressesResponse from(Address address) {
+        private static AddressesResponse from(Address address) {
             return new AddressesResponse(address.getId(), address.getName());
         }
     }
