+++ be/src/main/java/kr/codesquad/secondhand/api/member/dto/MemberAddressResponse.java
@@ -2,7 +2,7 @@ package kr.codesquad.secondhand.api.member.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kr.codesquad.secondhand.api.member.domain.Address;
+import kr.codesquad.secondhand.api.address.domain.Address;
 import kr.codesquad.secondhand.api.member.domain.MemberAddress;
 import lombok.Getter;
