+++ be/src/main/java/kr/codesquad/secondhand/api/product/service/ProductFacadeService.java
@@ -3,10 +3,10 @@ package kr.codesquad.secondhand.api.product.service;
 import java.io.IOException;
 import java.net.URL;
 import java.util.List;
+import kr.codesquad.secondhand.api.address.domain.Address;
+import kr.codesquad.secondhand.api.address.repository.AddressRepositoryImpl;
 import kr.codesquad.secondhand.api.category.domain.Category;
-import kr.codesquad.secondhand.api.member.domain.Address;
 import kr.codesquad.secondhand.api.member.domain.Member;
-import kr.codesquad.secondhand.api.member.repository.AddressRepositoryImpl;
 import kr.codesquad.secondhand.api.member.service.MemberService;
 import kr.codesquad.secondhand.api.product.domain.Product;
 import kr.codesquad.secondhand.api.product.domain.ProductImage;
